# [Merged by Bors] - perf: short circuit wait for metadata