--- a/crates/fluvio/src/sync/mod.rs
+++ b/crates/fluvio/src/sync/mod.rs
@@ -109,9 +109,8 @@ mod context {
             use fluvio_future::timer::sleep;
 
             // We can short circuit here if already present
-            let found = search(self.store().read().await);
-            if found.is_some() {
-                return Ok(found);
+            if let Some(found) = search(self.store().read().await) {
+                return Ok(Some(found));
             }
 
             let mut timer = sleep(Duration::from_millis(*MAX_WAIT_TIME));