1. Added some performance fixes to producer test framework itself.
2. Performance improvements to `sync::lookup_and_wait`
3. Benchmarked results for producer 1 X 100,000 and 3 X 150,000.

Example bench output
```
Start running fluvio test runner
Testing connection to Fluvio cluster in profile
starting test in child process
Creating the topic: producer-test
topic "producer-test" already exists

Starting Producer test
Producers: 3
# Records: 150000
Starting Producer #0
Starting Producer #1
Starting Producer #2
[producer-2] Produce P99: 6.783µs Peak Throughput: 155189 kB/s
[producer-1] Produce P99: 6.751µs Peak Throughput: 157286 kB/s
[producer-0] Produce P99: 6.783µs Peak Throughput: 157286 kB/s
test passed, signalling success to parent
Test passed
+----------+---------------+
| Pass?    | true          |
|----------+---------------|
| Duration | 23.135994663s |
+----------+---------------+
```


### profiling results
No change

```
# 1 X 100K
[producer-0] Produce P99: 18.175µs Peak Throughput: 65011 kB/s

#3 X 150
[producer-0] Produce P99: 19.199µs Peak Throughput: 62128 kB/s
[producer-1] Produce P99: 19.327µs Peak Throughput: 61341 kB/s
[producer-2] Produce P99: 19.199µs Peak Throughput: 61865 kB/s
```

Test harness improvements only
```
# 1 X 100K
[producer-0] Produce P99: 13.759µs Peak Throughput: 78118 kB/s

#3 X 150
[producer-2] Produce P99: 13.951µs Peak Throughput: 74973 kB/s
[producer-1] Produce P99: 14.015µs Peak Throughput: 76021 kB/s
[producer-0] Produce P99: 14.079µs Peak Throughput: 74448 kB/s
```

Short circuiting
```
# 1 X 100K
[producer-0] Produce P99: 7.103µs Peak Throughput: 147849 kB/s

#3 X 150
[producer-2] Produce P99: 6.783µs Peak Throughput: 155189 kB/s
[producer-1] Produce P99: 6.751µs Peak Throughput: 157286 kB/s
[producer-0] Produce P99: 6.783µs Peak Throughput: 157286 kB/s
```