@davidbeesley 2022-10-13
> Can you update benchmark with latest changes?

The benchmark didn't change at all because for all iterations except the first it short circuits and that part hasn't changed. I confirmed this by rerunning.