@davidbeesley 2022-10-13
> LGTM. Thanks for contribution and nice work!

Thanks. There is still one more loose end I need to tie up before this is ready to ship.
The consumer test used to use the `production-test` topic and I switched it to use `consumer-test`. However, it doesn't populate the topic with any records before running so right now the smoke-test is essentially a no-op. The test should also fail if it didn't get as many records at it expected.